services/api/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio>=5.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# LLM Providers
groq>=0.4.0
google-generativeai>=0.3.0

# Image handling
pillow>=10.2.0
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0  # For signing player tokens
ulid-py>=1.1.0  # For generating IDs

# Development
ruff>=0.1.0
pytest>=7.4.0
pytest-asyncio>=0.23.0

